\name{glide_plot}
\alias{glide_plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function used to plot
}
\description{
It is used to draw MR-Egger regression plot and q-q plot
}
\usage{
glide_plot(out,qcutoff)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{out}{
out is a dataframe, it is the outcome of glide function.
qcutoff is the q-value cutoff.
}
}
